adding jwt claims in delete items and jwt optional in get items

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource,reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import get_jwt, get_jwt_identity, jwt_required
 from models.item import ItemModel
 
 # resources
@@ -18,7 +18,7 @@ class Item(Resource):
     )
 
     # retrieves an item from the list that matches the name given
-    @jwt_required
+    @jwt_required()
     def get(self,name):
         item = ItemModel.find_by_name(name) # item object
         if item:
@@ -45,7 +45,12 @@ class Item(Resource):
         # 201 represents created
 
     # deleting an item
+    @jwt_required()
     def delete(self,name):
+        # checking if the logged in user is an admin or not
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message' : 'You need to be an admin.'}
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -66,5 +71,18 @@ class Item(Resource):
 
 class ItemList(Resource):
     # getting the list of items
+    @jwt_required(optional=True) # jwt is not required, so a user can access this even when it is logged out 
     def get(self):
-        return {'items' : [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity() # is user is not logged in then None will be returned
+        print(user_id)
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            print('logged in')
+            # user is logged in
+            return {'items' : items}
+        # user is not logged in 
+        print('logged out')
+        return {
+            'items' : [item['name'] for item in items],
+            'message': 'More information available if you log in.'
+            }
